Give float-typed variables float literals
`total: float = 0` and `ratio = math.floor(ratio)` are both fine to the
interpreter and to mypy, which widens int to float freely. A compiled
build does not: mypyc boxes the two differently and refuses the
assignment outright.

Nothing here changes a number. It is the difference between a variable
whose declared type and whose literal agree and one where they only
happen to be compatible, and it costs nothing to have them agree - which
is worth doing whether or not this project ever compiles anything, and
was found by trying to.

+++ src/chunksim/model/rates.py
@@ -206,7 +206,10 @@ def find_fraction(value: float, rounded_denominator: bool = False) -> str:
         return "NaN"
     ratio = _round_half_up(1 / value, 2)
     if rounded_denominator:
-        ratio = math.floor(ratio)
+        # `float(...)` because `ratio` is typed `float` and `math.floor`
+        # narrows to `int`: harmless to the interpreter, but a compiled build
+        # boxes the two differently and refuses the assignment.
+        ratio = float(math.floor(ratio))
     return f"1/{_format_js_number(ratio)}"
